rgw/restore: Mark the restore entry status as `None` first time
While adding the restore entry to the FIFO, mark its status as `None`
so that restore thread knows that the entry is being processed for
the first time. Incase the restore is still in progress and the entry
needs to be re-added to the queue, its status then will be marked
`InProgress`.

Signed-off-by: Soumya Koduri <skoduri@redhat.com>
(cherry picked from commit 1eb0623d6b7dee653c7529a38b8f907185107d33)